Structural Revisions
Simplified grid to and move specialized code to the Paint Screen; this
will help in the reusability of the Grid.
Working on the FrameManager; once completed the Paint Screen; will be
able to draw on the ActiveFrame. Also the FramesPalette in the Workspace
will be able to display each of the frames in the List<Frames>.