Updated Pencil, and Input()
The Input class is now updated on each cycle to distinguish the common
checks for Left Button Pressed, Released, Held, and Inactive, to keep
from having all of the checks against current state and previous state
throughout code.
The Pencil now checks to see if a movement greater that one "Grid Pixel"
has taken place while drawing with the Pencil.  If the a larger that one
Pixel has happened a line will be drawn from previous position to
current position instead of having the "dotted" lines from quick
movement.